add masterchain support

+++ address/addr.go
@@ -37,11 +37,11 @@ type flags struct {
 	testnet    bool
 }
 
-func NewAddress(flags byte, workchain byte, data []byte) *Address {
+func NewAddress(flags byte, workchain int, data []byte) *Address {
 	return &Address{
 		flags:     parseFlags(flags),
 		addrType:  StdAddress,
-		workchain: int32(int8(workchain)),
+		workchain: int32(workchain),
 		bitsLen:   256,
 		data:      data,
 	}
